requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
requests==2.32.5
PyPDF2==3.0.1
pdfplumber==0.11.4
python-dotenv==1.0.1
spacy==3.8.11
reportlab==4.0.9
google-generativeai==0.8.3
